Fix remaining files to not complain due to -Wtruncate. Fixes from Brian Callahan, slightly modified my me.

+++ arch/amd64/code.c
@@ -262,7 +262,7 @@ bfcode(struct symtab **s, int cnt)
 		if (sp == NULL)
 			continue; /* XXX when happens this? */
 
-		ssz = tsize(sp->stype, sp->sdf, sp->sap);
+		ssz = (int)tsize(sp->stype, sp->sdf, sp->sap);
 		switch (typ = argtyp(sp->stype, sp->sdf, sp->sap)) {
 		case INTEGER:
 		case SSE:
