+++ cc/cxxcom/symtabs.c
@@ -219,7 +219,7 @@ lookup(char *key, int stype)
 		w = sympole[type];
 		for (;;) {
 			bit = BITNO(w->bitno);
-			fbit = (code >> bit) & 1;
+			fbit = (int)(code >> bit) & 1;
 			svbit = fbit ? IS_RIGHT_LEAF(w->bitno) :
 			    IS_LEFT_LEAF(w->bitno);
 			w = w->lr[fbit];
@@ -267,7 +267,7 @@ lookup(char *key, int stype)
 
 	new = stype & STEMP ? tmpalloc(sizeof(struct tree)) :
 	    permalloc(sizeof(struct tree));
-	bit = (code >> cix) & 1;
+	bit = (int)(code >> cix) & 1;
 	new->bitno = cix | (bit ? RIGHT_IS_LEAF : LEFT_IS_LEAF);
 	new->lr[bit] = (struct tree *)getsymtab(key, stype);
 	if (numsyms[type]++ == 1) {
@@ -287,7 +287,7 @@ lookup(char *key, int stype)
 			cerror("bitno == cix");
 		if (bitno > cix)
 			break;
-		svbit = (code >> bitno) & 1;
+		svbit = (int)(code >> bitno) & 1;
 		last = w;
 		w = w->lr[svbit];
 		if (fbit & (svbit ? RIGHT_IS_LEAF : LEFT_IS_LEAF))
